Revert "separate tasks for export table data and store file as document"
This reverts commit 8c8566322540a163425ba58aea7f1dd766b81aba.
